Add DID resolution types and rename UniversalResolver
